Add files via upload
'AmideSearch_Strongylocentrotus pupuratus' is an excel file containing the information on proteins from an NCBI BLAST search with the motif xGyz up to 350 residues in length. x is any amino acid residue, y and z are arginine or lysine.
Recent Searches of interest_Strongylocentrotus purpuratus’ is an excel file containing uncharacterised proteins from 'AmideSearch_Strongylocentrotus pupuratus' that have a signal peptide probability over 0.5. 'The Already known NP' sheet contains a list of the known neuropeptides found in the amide search.